requirements_test.txt
-r requirements_dev.txt
pytest==7.1.3
pytest-homeassistant-custom-component==0.12.21
